+++ backend/src/main/java/com/bd2_team6/biteright/controllers/DTO/DailyLimitsDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DailyLimitsDTO {
-    private Integer dailyLimitId;
+    private Long dailyLimitId;
     private Integer calorieLimit;
     private Integer proteinLimit;
     private Integer fatLimit;
