+++ backend/src/main/java/com/bd2_team6/biteright/entities/daily_limits/DailyLimits.java
@@ -16,7 +16,7 @@ public class DailyLimits {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "daily_limit_id")
-    private Integer dailyLimitId;
+    private Long dailyLimitId;
 
     @Column(name = "calorie_limit")
     private Integer calorieLimit;
