+++ backend/src/main/java/com/bd2_team6/biteright/entities/meal/Meal.java
@@ -21,7 +21,7 @@ public class Meal {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "meal_id")
-    private Integer mealId;
+    private Long mealId;
 
     @Column(name = "meal_date")
     private LocalDateTime mealDate;
