+++ backend/src/main/java/com/bd2_team6/biteright/entities/user_info/UserInfoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface UserInfoRepository extends JpaRepository<UserInfo, Integer> {
+public interface UserInfoRepository extends JpaRepository<UserInfo, Long> {
     
 }
