+++ backend/src/main/java/com/bd2_team6/biteright/entities/user_preferences/UserPreferencesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface UserPreferencesRepository extends JpaRepository<UserPreferences, Integer> {
+public interface UserPreferencesRepository extends JpaRepository<UserPreferences, Long> {
     
 }
